#105 #104 #96 #94 #39 Refactor client-server communication
-Update dependencies to conform to sprotty 0.9.0 and Theia 1.3.0
- Adapt code to confrom to client/server changes
- Change argument for passing the server port to "WF_GLSP"

Requires: 
- eclipse-glsp/glsp-server/pull/73
- eclipse-glsp/glsp-theia-integration/pull/44

Part of:
- eclipse-glsp/glsp/issues/104
- eclipse-glsp/glsp/issues/94
- eclipse-glsp/glsp/issues/96
- eclipse-glsp/glsp/issues/105
