+++ client/package.json
@@ -25,24 +25,28 @@
     "lerna": "^2.2.0",
     "tslint": "^5.5.0",
     "tslint-loader": "^3.4.3",
-    "typescript": "3.6.4"
+    "typescript": "^3.9.2"
   },
   "resolutions": {
-    "**/@theia/application-manager": "1.0.0",
-    "**/@theia/application-package": "1.0.0",
-    "**/@theia/core": "1.0.0",
-    "**/@theia/editor": "1.0.0",
-    "**/@theia/filesystem": "1.0.0",
-    "**/@theia/languages": "1.0.0",
-    "**/@theia/markers": "1.0.0",
-    "**/@theia/messages": "1.0.0",
-    "**/@theia/monaco": "1.0.0",
-    "**/@theia/navigator": "1.0.0",
-    "**/@theia/output": "1.0.0",
-    "**/@theia/preferences": "1.0.0",
-    "**/@theia/process": "1.0.0",
-    "**/@theia/terminal": "1.0.0",
-    "**/@theia/workspace": "1.0.0"
+    "**/@theia/application-manager": "1.3.0",
+    "**/@theia/application-package": "1.3.0",
+    "**/@theia/core": "1.3.0",
+    "**/@theia/editor": "1.3.0",
+    "**/@theia/filesystem": "1.3.0",
+    "**/@theia/languages": "1.3.0",
+    "**/@theia/markers": "1.3.0",
+    "**/@theia/messages": "1.3.0",
+    "**/@theia/monaco": "1.3.0",
+    "**/@theia/navigator": "1.3.0",
+    "**/@theia/output": "1.3.0",
+    "**/@theia/outline-view": "1.3.0",
+    "**/@theia/preferences": "1.3.0",
+    "**/@theia/variable-resolver": "1.3.0",
+    "**/@theia/process": "1.3.0",
+    "**/@theia/terminal": "1.3.0",
+    "**/@theia/workspace": "1.3.0",
+    "**/sprotty": "0.9.0",
+    "**/sprotty-theia": "0.9.0"
   },
   "workspaces": [
     "workflow/*"
