+++ client/workflow/browser-app/package.json
@@ -3,27 +3,27 @@
   "name": "browser-app",
   "version": "0.1.0",
   "dependencies": {
-    "@theia/core": "1.0.0",
-    "@theia/editor": "1.0.0",
-    "@theia/filesystem": "1.0.0",
-    "@theia/languages": "1.0.0",
-    "@theia/markers": "1.0.0",
-    "@theia/messages": "1.0.0",
-    "@theia/monaco": "1.0.0",
-    "@theia/navigator": "1.0.0",
-    "@theia/preferences": "1.0.0",
-    "@theia/process": "1.0.0",
-    "@theia/terminal": "1.0.0",
-    "@theia/workspace": "1.0.0",
+    "@theia/core": "1.3.0",
+    "@theia/editor": "1.3.0",
+    "@theia/filesystem": "1.3.0",
+    "@theia/languages": "1.3.0",
+    "@theia/markers": "1.3.0",
+    "@theia/messages": "1.3.0",
+    "@theia/monaco": "1.3.0",
+    "@theia/navigator": "1.3.0",
+    "@theia/preferences": "1.3.0",
+    "@theia/process": "1.3.0",
+    "@theia/terminal": "1.3.0",
+    "@theia/workspace": "1.3.0",
     "@eclipse-glsp-examples/workflow-theia": "0.7.1"
   },
   "devDependencies": {
-    "@theia/cli": "1.0.0"
+    "@theia/cli": "1.3.0"
   },
   "scripts": {
     "prepare": "theia build --mode development",
-    "start": "theia start --WORKFLOW_LSP=5007 --root-dir=../workspace",
-    "start:debug": "theia start --WORKFLOW_LSP=5007  --root-dir=../workspace --loglevel=debug",
+    "start": "theia start --WF_GLSP=5007 --root-dir=../workspace",
+    "start:debug": "theia start --WF_GLSP=5007  --root-dir=../workspace --loglevel=debug",
     "watch": "theia build --watch --mode development"
   },
   "theia": {
