+++ client/workflow/workflow-sprotty/package.json
@@ -16,7 +16,7 @@
   "devDependencies": {
     "rimraf": "^2.6.1",
     "tslint": "^5.5.0",
-    "typescript": "3.4.3"
+    "typescript": "^3.9.2"
   },
   "scripts": {
     "prepare": "yarn run clean && yarn run build",
