+++ client/workflow/workflow-theia/src/node/workflow-glsp-server-contribution.ts
@@ -13,22 +13,12 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
-import { BaseGLSPServerContribution } from "@eclipse-glsp/theia-integration/lib/node";
-import { IConnection } from "@theia/languages/lib/node";
+import { BaseGLSPServerContribution, getPort } from "@eclipse-glsp/theia-integration/lib/node";
 import { injectable } from "inversify";
 import * as net from "net";
-import { createSocketConnection } from "vscode-ws-jsonrpc/lib/server";
+import { createSocketConnection, IConnection } from "vscode-ws-jsonrpc/lib/server";
 
 import { WorkflowLanguage } from "../common/workflow-language";
-
-function getPort(): number | undefined {
-    const arg = process.argv.filter(arg => arg.startsWith('--WORKFLOW_LSP='))[0];
-    if (!arg) {
-        return undefined;
-    } else {
-        return Number.parseInt(arg.substring('--WORKFLOW_LSP='.length), 10);
-    }
-}
 
 @injectable()
 export class WorkflowGLServerContribution extends BaseGLSPServerContribution {
@@ -45,8 +35,8 @@ export class WorkflowGLServerContribution extends BaseGLSPServerContribution {
     };
 
     start(clientConnection: IConnection): void {
-        const socketPort = getPort();
-        if (socketPort) {
+        const socketPort = getPort("WF_GLSP");
+        if (!isNaN(socketPort)) {
             const socket = new net.Socket();
             const serverConnection = createSocketConnection(socket, socket, () => {
                 socket.destroy();
