+++ server/org.eclipse.glsp.example.workflow/src/main/java/org/eclipse/glsp/example/workflow/WorkflowGLSPServer.java
@@ -15,8 +15,6 @@
  ********************************************************************************/
 package org.eclipse.glsp.example.workflow;
 
-import java.util.concurrent.CompletableFuture;
-
 import org.apache.log4j.Logger;
 import org.eclipse.glsp.server.jsonrpc.DefaultGLSPServer;
 
@@ -28,10 +26,10 @@ public class WorkflowGLSPServer extends DefaultGLSPServer<WorkflowInitializeOpti
    }
 
    @Override
-   public CompletableFuture<Boolean> handleOptions(final WorkflowInitializeOptions options) {
+   public boolean handleOptions(final WorkflowInitializeOptions options) {
       if (options != null) {
          LOGGER.debug(options.getTimestamp() + ": " + options.getMessage());
       }
-      return CompletableFuture.completedFuture(true);
+      return true;
    }
 }
