+++ server/org.eclipse.glsp.example.workflow/src/main/java/org/eclipse/glsp/example/workflow/taskedit/EditTaskOperationHandler.java
@@ -17,8 +17,8 @@ package org.eclipse.glsp.example.workflow.taskedit;
 
 import java.util.Optional;
 
-import org.eclipse.glsp.api.jsonrpc.GLSPServerException;
 import org.eclipse.glsp.api.model.GraphicalModelState;
+import org.eclipse.glsp.api.protocol.GLSPServerException;
 import org.eclipse.glsp.example.workflow.wfgraph.TaskNode;
 import org.eclipse.glsp.server.operationhandler.BasicOperationHandler;
